executor: move some exec out of executor.go
